+++ index.html
@@ -8,17 +8,8 @@
 </head>
 
 <body>
-  <div class="container">
-    <ul class="comments">
-      <!-- рендеринг -->
-    </ul>
-    <div class="add-form">
-      <input type="text" class="add-form-name" placeholder="Введите ваше имя" />
-      <textarea type="textarea" class="add-form-text" placeholder="Введите ваш коментарий" rows="4"></textarea>
-      <div class="add-form-row">
-        <button class="add-form-button">Написать</button>
-      </div>
-    </div>
+  <div id="app">
+
   </div>
 </body>
 
@@ -28,8 +19,8 @@
   }
 </style>
 
-<script type="module" src="script.js"></script>
-<script type="module" src="api.js"></script>
-<script type="module" src="render.js"></script>
+<script src="script.js"></script>
+<!-- <script type="module" src="api.js"></script>
+<script type="module" src="render.js"></script> -->
 
 </html>
