+++ styles.css
@@ -22,7 +22,9 @@ body {
   }
 
   .comment,
-  .add-form {
+  .add-form,
+  .login-form,
+  .reg-form {
     width: 596px;
     box-sizing: border-box;
     background: radial-gradient(
@@ -91,7 +93,9 @@ body {
     background-image: url("data:image/svg+xml,%3Csvg width='22' height='20' viewBox='0 0 22 20' fill='none' xmlns='http://www.w3.org/2000/svg'%3E%3Cpath d='M15.95 0C14.036 0 12.199 0.882834 11 2.26703C9.801 0.882834 7.964 0 6.05 0C2.662 0 0 2.6267 0 5.99455C0 10.1035 3.74 13.4714 9.405 18.5613L11 20L12.595 18.5613C18.26 13.4714 22 10.1035 22 5.99455C22 2.6267 19.338 0 15.95 0Z' fill='%23BCEC30'/%3E%3C/svg%3E");
   }
 
-  .add-form {
+  .add-form,
+  .login-form,
+  .reg-form {
     padding: 20px;
     margin-top: 48px;
     display: flex;
@@ -99,14 +103,24 @@ body {
   }
 
   .add-form-name,
-  .add-form-text {
+  .add-form-text,
+  .login-name,
+  .login-password,
+  .reg-name,
+  .reg-login,
+  .reg-password {
     font-size: 16px;
     font-family: Helvetica;
     border-radius: 8px;
     border: none;
   }
 
-  .add-form-name {
+  .add-form-name,
+  .login-name,
+  .login-password,
+  .reg-name,
+  .reg-login,
+  .reg-password {
     width: 300px;
     padding: 11px 22px;
   }
@@ -122,7 +136,9 @@ body {
     justify-content: flex-end;
   }
 
-  .add-form-button {
+  .add-form-button,
+  .login-button,
+  .reg-button {
     margin-top: 24px;
     font-size: 24px;
     padding: 10px 20px;
@@ -134,4 +150,16 @@ body {
 
   .add-form-button:hover {
     opacity: 0.9;
+  }
+
+  .login-button:hover {
+    opacity: 0.9;
+  }
+
+  .reg-button:hover {
+    opacity: 0.9;
+  }
+
+  .link {
+    color: white;
   }
